Update bot.py
Added the option to start a new conversation. Please note if you want to use an old conversation please specify the parentid as well and vice versa! By default, a new conversation is not started and the bot will continue using the first conversation created.
Also fixed the refresh command to actually run.